+++ src/main/java/com/mapu/global/jwt/exception/errorcode/JwtExceptionErrorCode.java
@@ -9,7 +9,7 @@ public enum JwtExceptionErrorCode implements ResponseStatus {
      * 7000: JWT 오류
      */
     ERROR_IN_JWT(7000, HttpStatus.INTERNAL_SERVER_ERROR.value(), "JWT 토큰 발급 중 오류가 발생했습니다."),
-    NO_JWT_TOKEN_IN_HEADER(7001, HttpStatus.UNAUTHORIZED.value(), "헤더에 JWT 토큰이 없습니다."),
+    NO_JWT_TOKEN(7001, HttpStatus.UNAUTHORIZED.value(), "JWT 토큰이 없습니다."),
     NO_BEARER_TYPE(7002, HttpStatus.UNAUTHORIZED.value(),"Bearer 토큰이 아닙니다."),
     EXPIRED_JWT_TOKEN(7003, HttpStatus.UNAUTHORIZED.value(), "만료된 JWT 토큰입니다."),
     WRONG_JWT_TOKEN_TYPE(7004, HttpStatus.UNAUTHORIZED.value(), "JWT 토큰 타입이 올바르지 않습니다."),
